Concurrent Dictionary for the loggers
