+++ src/JJMasterData.Commons/Logging/File/FileLogger.cs
@@ -5,19 +5,26 @@ using System.Text;
 using System.Threading.Tasks;
 using JJMasterData.Commons.Util;
 using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
 
 namespace JJMasterData.Commons.Logging.File;
 
 public class FileLogger : ILogger
 {
-    private readonly FileLoggerProvider _fileLoggerProvider;
+
+    private readonly string _name;
+    private readonly BlockingCollection<LogMessage> _queue;
+    private readonly Func<FileLoggerOptions> _getCurrentOptions;
 
     /// <summary>
     /// Creates a new instance of <see cref="FileLogger" />.
     /// </summary>
-    public FileLogger(FileLoggerProvider fileLoggerProvider)
+    public FileLogger(string name, BlockingCollection<LogMessage> queue, Func<FileLoggerOptions> getCurrentOptions)
     {
-        _fileLoggerProvider = fileLoggerProvider;
+        _name = name;
+        _queue = queue;
+        _getCurrentOptions = getCurrentOptions;
+        Task.Factory.StartNew(LogAtFile, TaskCreationOptions.LongRunning);
     }
 
     public IDisposable BeginScope<TState>(TState state) => default!;
@@ -47,6 +54,90 @@ public class FileLogger : ILogger
             return;
 
         var message = new LogMessage(logLevel, eventId, state!, exception, (s, e) => formatter((TState)s, e));
-        _fileLoggerProvider.AddToQueue(message);
+        _queue.Add(message);
+    }
+    
+    private void LogAtFile()
+    {
+        var options = _getCurrentOptions();
+        foreach (var message in _queue.GetConsumingEnumerable())
+        {
+            var path = FileIO.ResolveFilePath(options.FileName);
+            var directory = Path.GetDirectoryName(path);
+
+            if (!Directory.Exists(directory))
+                Directory.CreateDirectory(directory!);
+
+            var formatting = options.Formatting;
+            var record = GetLogRecord(message, formatting);
+            
+            using var writer = new StreamWriter(path, true);
+            writer.Write(record);
+        }
+    }
+
+    private static string GetLogRecord(LogMessage message, FileLoggerFormatting formatting)
+    {
+        var log = new StringBuilder();
+
+        switch (formatting)
+        {
+            case FileLoggerFormatting.Default:
+
+                log.Append(DateTime.Now);
+                log.Append(" ");
+                if (!string.IsNullOrWhiteSpace(message.EventId.Name))
+                {
+                    log.AppendFormat(" [{0}] ", message.EventId.Name);
+                }
+
+                log.Append("(");
+                log.Append(message.LogLevel.ToString());
+                log.AppendLine(")");
+                log.Append(message.Formatter(message.State, message.Exception));
+                log.AppendLine();
+                log.AppendLine();
+                break;
+            case FileLoggerFormatting.Compact:
+            {
+                log.AppendFormat("{0:yyyy-MM-dd HH:mm:ss+00:00} -", DateTime.Now);
+                log.AppendFormat(" [{0}] ", message.LogLevel);
+
+                if (!string.IsNullOrWhiteSpace(message.EventId.Name))
+                {
+                    log.AppendFormat(" [{0}] ", message.EventId.Name);
+                }
+
+                log.AppendFormat(" {0} ", message.Formatter(message.State, message.Exception));
+
+                if (message.Exception != null)
+                {
+                    log.AppendLine(message.Exception.Message);
+                    log.AppendLine(message.Exception.StackTrace);
+                    log.AppendFormat("Source: {0}", message.Exception.Source);
+                }
+                log.AppendLine();
+                break;
+            }
+            case FileLoggerFormatting.Json:
+                log.Append(
+                    JsonConvert.SerializeObject(new
+                    {   Date = DateTime.Now,
+                        Event = message.EventId.Name,
+                        message.LogLevel,
+                        Message = message.Formatter(message.State, message.Exception),
+                        message.Exception
+                    }, new JsonSerializerSettings()
+                    {
+                        NullValueHandling = NullValueHandling.Ignore,
+                        Formatting = Formatting.Indented
+                    }));
+                log.AppendLine(",");
+                break;
+            default:
+                throw new ArgumentOutOfRangeException(nameof(formatting), formatting, null);
+        }
+
+        return log.ToString();
     }
 }
